requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
pydantic==2.10.3
pydantic-settings==2.7.0
jinja2==3.1.4
tavily-python==0.5.0
python-dotenv==1.0.1
httpx==0.27.2
sqlalchemy==2.0.36
alembic==1.14.0
psycopg2-binary==2.9.10
asyncpg==0.30.0
groq>=0.9.0
